odep/metamind/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metamind"
version = "4.0.0"
description = "MetaMind Enterprise Query Intelligence Platform"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "MetaMind Team", email = "team@metamind.io"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    
    # SQL Parsing
    "sqlglot>=20.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "oracledb>=1.4.0",
    "redis>=5.0.0",
    "aioredis>=2.0.0",
    
    # Data Processing
    "pyarrow>=14.0.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    
    # ML/AI
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "torch>=2.1.0",
    
    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Observability
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    
    # Utilities
    "structlog>=23.2.0",
    "tenacity>=8.2.0",
    "python-json-logger>=2.0.7",
    "typer>=0.9.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
gpu = [
    "cudf-cu12>=23.10.0",
    "cuml-cu12>=23.10.0",
    "cugraph-cu12>=23.10.0",
    "pylibraft-cu12>=23.10.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "httpx>=0.25.0",
    "factory-boy>=3.3.0",
    "faker>=20.1.0",
]

[project.scripts]
metamind = "metamind.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["metamind*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "gpu: GPU tests",
]

[tool.coverage.run]
source = ["metamind"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
show_missing = true
fail_under = 80
